refactor: move Settings to engine
