+++ core/CMakeLists.txt
@@ -25,7 +25,6 @@ message("# Building core tests: " ${BUILD_CORE_TESTS})
 
 set(CUBOS_CORE_SOURCE
     "src/cubos/core/log.cpp"
-    "src/cubos/core/settings.cpp"
     "src/cubos/core/thread_pool.cpp"
 
     "src/cubos/core/memory/stream.cpp"
@@ -88,7 +87,6 @@ set(CUBOS_CORE_SOURCE
 
 set(CUBOS_CORE_INCLUDE
     "include/cubos/core/log.hpp"
-    "include/cubos/core/settings.hpp"
     "include/cubos/core/thread_pool.hpp"
 
     "include/cubos/core/memory/move.hpp"
