+++ engine/include/cubos/engine/window/plugin.hpp
@@ -37,6 +37,9 @@ namespace cubos::engine
     /// ## Tags
     /// - `cubos.window.poll` - the window is polled for events, sending @ref core::io::WindowEvent's.
     /// - `cubos.window.render` - the window's back buffers are swapped.
+    ///
+    /// ## Dependencies
+    /// - @ref settings-plugin
 
     /// @brief Plugin entry function.
     /// @param cubos @b CUBOS. main class
