+++ engine/samples/assets/json/main.cpp
@@ -11,18 +11,18 @@
 #include <vector>
 
 #include <cubos/core/data/fs/file_system.hpp>
-#include <cubos/core/settings.hpp>
 
 #include <cubos/engine/assets/bridges/json.hpp>
 #include <cubos/engine/assets/plugin.hpp>
+#include <cubos/engine/settings/settings.hpp>
 
-using cubos::core::Settings;
 using cubos::core::data::Deserializer;
 using cubos::core::data::FileSystem;
 using cubos::core::data::Serializer;
 using cubos::core::ecs::Read;
 using cubos::core::ecs::Write;
 using cubos::core::memory::Stream;
+
 using namespace cubos::engine;
 
 /// A simple serializable type which we will be saving and loading.
