+++ engine/src/cubos/engine/renderer/plugin.cpp
@@ -3,7 +3,6 @@
 
 #include <cubos/core/ecs/query.hpp>
 #include <cubos/core/gl/camera.hpp>
-#include <cubos/core/settings.hpp>
 
 #include <cubos/engine/renderer/deferred_renderer.hpp>
 #include <cubos/engine/renderer/environment.hpp>
@@ -11,10 +10,10 @@
 #include <cubos/engine/renderer/light.hpp>
 #include <cubos/engine/renderer/plugin.hpp>
 #include <cubos/engine/renderer/pps/bloom.hpp>
+#include <cubos/engine/settings/plugin.hpp>
 #include <cubos/engine/transform/plugin.hpp>
 #include <cubos/engine/window/plugin.hpp>
 
-using cubos::core::Settings;
 using cubos::core::ecs::EventReader;
 using cubos::core::ecs::Query;
 using cubos::core::ecs::Read;
@@ -22,6 +21,7 @@ using cubos::core::ecs::Write;
 using cubos::core::io::ResizeEvent;
 using cubos::core::io::Window;
 using cubos::core::io::WindowEvent;
+
 using namespace cubos::engine;
 
 static void init(Write<Renderer> renderer, Read<Window> window, Write<Settings> settings)
